+++ src/CMakeLists.txt
@@ -31,6 +31,7 @@ endif()
 
 # Compile src/lua/*.lua files into src/lua/*.lua.c sources
 set(lua_sources)
+lua_source(lua_sources lua/loaders.lua loaders_lua)
 lua_source(lua_sources lua/init.lua init_lua)
 lua_source(lua_sources lua/debug.lua debug_lua)
 lua_source(lua_sources lua/string.lua string_lua)
